Move build command into own file
